Move networking classes, disconnect mismatched versions and include commit hash in dev versions
