+++ gradle.properties
@@ -9,7 +9,7 @@ mappings_channel    = official
 mappings_version    = 1.16.5
 
 # mod properties
-mod_version=3.0.0-dev-1.0
+mod_version=3.0.0-SNAPSHOT
 maven_group=net.splatcraft.forge
 archives_base_name=splatcraft
