+++ src/main/java/net/splatcraft/forge/Splatcraft.java
@@ -4,15 +4,16 @@ import net.minecraft.block.Block;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.RegistryEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
-import net.minecraftforge.fml.event.server.FMLServerAboutToStartEvent;
 import net.minecraftforge.fml.event.server.FMLServerStartedEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import net.minecraftforge.fml.loading.FMLPaths;
+import net.minecraftforge.fml.loading.moddiscovery.ModInfo;
 import net.splatcraft.forge.data.SplatcraftTags;
 import net.splatcraft.forge.handlers.ScoreboardHandler;
 import net.splatcraft.forge.handlers.client.ClientSetupHandler;
@@ -27,22 +28,23 @@ import net.splatcraft.forge.registries.SplatcraftItems;
 import net.splatcraft.forge.registries.SplatcraftStats;
 import net.splatcraft.forge.registries.SplatcraftTileEntitites;
 import net.splatcraft.forge.world.gen.SplatcraftOreGen;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
+
+import java.util.Objects;
 
 // The value here should match an entry in the META-INF/mods.toml file
 @Mod(Splatcraft.MODID)
-public class Splatcraft
-{
+public class Splatcraft {
     public static final String MODID = "splatcraft";
-    public static final String MODNAME = "Splatcraft";
-    public static final String SHORT = "SC";
-    public static final String VERSION = "2.3.0";
+    public static String version;
     // Directly reference a log4j logger.
-    private static final Logger LOGGER = LogManager.getLogger();
 
-    public Splatcraft()
-    {
+    public Splatcraft() {
+        for (ModInfo m : ModList.get().getMods()) { // Forge is stupid
+            if (Objects.equals(m.getModId(), MODID) && m.getVersion() != null) {
+                version = m.getVersion().toString();
+                break;
+            }
+        }
         ModLoadingContext.get().registerConfig(ModConfig.Type.CLIENT, SplatcraftConfig.clientConfig);
         SplatcraftConfig.loadConfig(SplatcraftConfig.clientConfig, FMLPaths.CONFIGDIR.get().resolve(Splatcraft.MODID + "-client.toml").toString());
 
@@ -72,8 +74,7 @@ public class Splatcraft
         SplatcraftOreGen.registerOres();
     }
 
-    private void clientSetup(final FMLClientSetupEvent event)
-    {
+    private void clientSetup(final FMLClientSetupEvent event) {
         SplatcraftEntities.bindRenderers();
         SplatcraftKeyHandler.registerKeys();
         SplatcraftBlocks.setRenderLayers();
@@ -86,12 +87,6 @@ public class Splatcraft
             SplatcraftItems.registerArmorModels();
             ClientSetupHandler.bindScreenContainers();
         });
-    }
-
-    @SubscribeEvent
-    public void onServerAboutToStart(FMLServerAboutToStartEvent event)
-    {
-
     }
 
     @SubscribeEvent
