+++ src/main/java/net/splatcraft/forge/handlers/SplatcraftCommonHandler.java
@@ -33,12 +33,12 @@ import net.splatcraft.forge.data.capabilities.playerinfo.IPlayerInfo;
 import net.splatcraft.forge.data.capabilities.playerinfo.PlayerInfoCapability;
 import net.splatcraft.forge.items.InkTankItem;
 import net.splatcraft.forge.items.InkWaxerItem;
-import net.splatcraft.forge.network.RequestPlayerInfoPacket;
 import net.splatcraft.forge.network.SplatcraftPacketHandler;
-import net.splatcraft.forge.network.UpdateBooleanGamerulesPacket;
-import net.splatcraft.forge.network.UpdateClientColorsPacket;
-import net.splatcraft.forge.network.UpdateColorScoresPacket;
-import net.splatcraft.forge.network.UpdatePlayerInfoPacket;
+import net.splatcraft.forge.network.c2s.RequestPlayerInfoPacket;
+import net.splatcraft.forge.network.s2c.UpdateBooleanGamerulesPacket;
+import net.splatcraft.forge.network.s2c.UpdateClientColorsPacket;
+import net.splatcraft.forge.network.s2c.UpdateColorScoresPacket;
+import net.splatcraft.forge.network.s2c.UpdatePlayerInfoPacket;
 import net.splatcraft.forge.registries.SplatcraftGameRules;
 import net.splatcraft.forge.tileentities.InkedBlockTileEntity;
 import net.splatcraft.forge.util.ColorUtils;
