+++ src/main/java/net/splatcraft/forge/network/c2s/DodgeRollPacket.java
@@ -1,10 +1,9 @@
-package net.splatcraft.forge.network;
+package net.splatcraft.forge.network.c2s;
 
-import net.splatcraft.forge.items.weapons.DualieItem;
-import net.splatcraft.forge.network.base.PlayToServerPacket;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.network.PacketBuffer;
+import net.splatcraft.forge.items.weapons.DualieItem;
 
 import java.util.UUID;
