+++ src/main/java/net/splatcraft/forge/network/c2s/PlayToServerPacket.java
@@ -1,8 +1,9 @@
-package net.splatcraft.forge.network.base;
+package net.splatcraft.forge.network.c2s;
 
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraftforge.fml.network.NetworkDirection;
 import net.minecraftforge.fml.network.NetworkEvent;
+import net.splatcraft.forge.network.SplatcraftPacket;
 
 import java.util.function.Supplier;
