+++ src/main/java/net/splatcraft/forge/network/c2s/RequestPlayerInfoPacket.java
@@ -1,9 +1,10 @@
-package net.splatcraft.forge.network;
+package net.splatcraft.forge.network.c2s;
 
-import net.splatcraft.forge.network.base.PlayToServerPacket;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.player.ServerPlayerEntity;
 import net.minecraft.network.PacketBuffer;
+import net.splatcraft.forge.network.SplatcraftPacketHandler;
+import net.splatcraft.forge.network.s2c.UpdatePlayerInfoPacket;
 
 import java.util.UUID;
