+++ src/main/java/net/splatcraft/forge/network/c2s/SwapSlotWithOffhandPacket.java
@@ -1,6 +1,5 @@
-package net.splatcraft.forge.network;
+package net.splatcraft.forge.network.c2s;
 
-import net.splatcraft.forge.network.base.PlayToServerPacket;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.network.PacketBuffer;
