+++ src/main/java/net/splatcraft/forge/network/s2c/PlayToClientPacket.java
@@ -1,7 +1,8 @@
-package net.splatcraft.forge.network.base;
+package net.splatcraft.forge.network.s2c;
 
 import net.minecraftforge.fml.network.NetworkDirection;
 import net.minecraftforge.fml.network.NetworkEvent;
+import net.splatcraft.forge.network.SplatcraftPacket;
 
 import java.util.function.Supplier;
