+++ src/main/java/net/splatcraft/forge/network/s2c/UpdateBooleanGamerulesPacket.java
@@ -1,9 +1,8 @@
-package net.splatcraft.forge.network;
+package net.splatcraft.forge.network.s2c;
 
-import net.splatcraft.forge.network.base.PlayToClientPacket;
-import net.splatcraft.forge.registries.SplatcraftGameRules;
 import net.minecraft.network.PacketBuffer;
 import net.minecraft.world.GameRules;
+import net.splatcraft.forge.registries.SplatcraftGameRules;
 
 import java.util.Map;
 import java.util.Set;
