+++ src/main/java/net/splatcraft/forge/util/ColorUtils.java
@@ -1,17 +1,5 @@
 package net.splatcraft.forge.util;
 
-import net.splatcraft.forge.blocks.IColoredBlock;
-import net.splatcraft.forge.client.particles.InkSplashParticleData;
-import net.splatcraft.forge.data.SplatcraftTags;
-import net.splatcraft.forge.data.capabilities.playerinfo.PlayerInfoCapability;
-import net.splatcraft.forge.entities.IColoredEntity;
-import net.splatcraft.forge.handlers.ScoreboardHandler;
-import net.splatcraft.forge.items.ColoredBlockItem;
-import net.splatcraft.forge.network.PlayerColorPacket;
-import net.splatcraft.forge.network.SplatcraftPacketHandler;
-import net.splatcraft.forge.registries.SplatcraftGameRules;
-import net.splatcraft.forge.registries.SplatcraftInkColors;
-import net.splatcraft.forge.tileentities.InkColorTileEntity;
 import net.minecraft.client.Minecraft;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.LivingEntity;
@@ -22,11 +10,28 @@ import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.vector.Vector3d;
-import net.minecraft.util.text.*;
+import net.minecraft.util.text.Color;
+import net.minecraft.util.text.ITextComponent;
+import net.minecraft.util.text.StringTextComponent;
+import net.minecraft.util.text.TextComponent;
+import net.minecraft.util.text.TextFormatting;
+import net.minecraft.util.text.TranslationTextComponent;
 import net.minecraft.world.World;
 import net.minecraft.world.server.ServerWorld;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
+import net.splatcraft.forge.blocks.IColoredBlock;
+import net.splatcraft.forge.client.particles.InkSplashParticleData;
+import net.splatcraft.forge.data.SplatcraftTags;
+import net.splatcraft.forge.data.capabilities.playerinfo.PlayerInfoCapability;
+import net.splatcraft.forge.entities.IColoredEntity;
+import net.splatcraft.forge.handlers.ScoreboardHandler;
+import net.splatcraft.forge.items.ColoredBlockItem;
+import net.splatcraft.forge.network.SplatcraftPacketHandler;
+import net.splatcraft.forge.network.s2c.PlayerColorPacket;
+import net.splatcraft.forge.registries.SplatcraftGameRules;
+import net.splatcraft.forge.registries.SplatcraftInkColors;
+import net.splatcraft.forge.tileentities.InkColorTileEntity;
 
 import java.util.Objects;
 import java.util.Random;
