+++ src/main/resources/META-INF/mods.toml
@@ -9,15 +9,15 @@ description     = '''
         Splatcraft is a Minecraft mod based on Nintendo's popular game Splatoon!
         This mod aims to add items and mechanics based on the game in a cohesive way, such as Turf Wars, Weapons, Squid Transformations and many more!
     '''
-authors         = "Cibernet, Andante"
-credits         = "Thanks to Nintendo and Mojang, for making such great games!"
-logoFile        = "pack.png"
+authors = "Cibernet, Andante"
+credits = "Thanks to Nintendo and Mojang, for making such great games!"
+logoFile = "pack.png"
 
-version         = "${file.jarVersion}"
+version = "${file.jarVersion}"
 # updateJSONURL   = "https://splatcraft.net/data/update/splatcraft.json"
 
-displayURL      = "https://curseforge.com/minecraft/mc-mods/splatcraft"
-issueTrackerURL = "https://github.com/splatcraft/splatcraft/issues"
+displayURL = "https://curseforge.com/minecraft/mc-mods/splatcraft"
+issueTrackerURL = "https://github.com/splatcraft/splatcraft-forge/issues"
 
 [[dependencies.splatcraft]]
 modId           = "forge"
